Refactor VS/VSR resources
