+++ tests/suite/test_app_protect_waf_policies.py
@@ -12,12 +12,16 @@ from suite.resources_utils import (
 from suite.custom_resources_utils import (
     create_crd_from_yaml,
     delete_crd,
+)
+from suite.vs_vsr_resources_utils import(
     delete_virtual_server,
     create_virtual_server_from_yaml,
     patch_virtual_server_from_yaml,
     patch_v_s_route_from_yaml,
     create_v_s_route_from_yaml,
     delete_v_s_route,
+)
+from suite.policy_resources_utils import(
     create_policy_from_yaml,
     delete_policy,
     read_policy,
