+++ tests/suite/test_batch_startup_times.py
@@ -33,14 +33,16 @@ from suite.resources_utils import (
     get_last_reload_status,
     get_pods_amount, get_total_vsr,
 )
-from suite.custom_resources_utils import (
+from suite.vs_vsr_resources_utils import (
     create_virtual_server_from_yaml,
     delete_virtual_server,
     patch_virtual_server_from_yaml,
-    create_policy_from_yaml,
-    delete_policy,
     create_virtual_server,
     create_v_s_route,
+)
+from suite.policy_resources_utils import (
+    create_policy_from_yaml,
+    delete_policy,
 )
 from suite.yaml_utils import get_first_ingress_host_from_yaml
 from settings import TEST_DATA
