+++ tests/suite/test_ingress_mtls.py
@@ -7,13 +7,15 @@ from suite.resources_utils import (
     delete_secret,
 )
 from suite.ssl_utils import create_sni_session
-from suite.custom_resources_utils import (
+from suite.vs_vsr_resources_utils import (
     read_vs,
     read_vsr,
     patch_virtual_server_from_yaml,
+    patch_v_s_route_from_yaml,
+)
+from suite.policy_resources_utils import (
     create_policy_from_yaml,
     delete_policy,
-    patch_v_s_route_from_yaml,
 )
 from settings import TEST_DATA
