+++ tests/suite/test_jwt_policies_vsr.py
@@ -9,14 +9,16 @@ from suite.resources_utils import (
 )
 from suite.custom_resources_utils import (
     read_custom_resource,
-    delete_virtual_server,
-    patch_virtual_server_from_yaml,
-    patch_v_s_route_from_yaml,
 )
 from suite.policy_resources_utils import (
     create_policy_from_yaml,
     delete_policy,
     read_policy,
+)
+from suite.vs_vsr_resources_utils import (
+    delete_virtual_server,
+    patch_virtual_server_from_yaml,
+    patch_v_s_route_from_yaml,
 )
 from settings import TEST_DATA, DEPLOYMENTS
