+++ tests/suite/test_v_s_route.py
@@ -5,7 +5,7 @@ from kubernetes.client.rest import ApiException
 from settings import TEST_DATA
 from suite.custom_assertions import assert_event_and_count, assert_event_and_get_count, \
     assert_event_with_full_equality_and_count
-from suite.custom_resources_utils import create_virtual_server_from_yaml, \
+from suite.vs_vsr_resources_utils import create_virtual_server_from_yaml, \
     delete_virtual_server, create_v_s_route_from_yaml, delete_v_s_route, get_vs_nginx_template_conf, \
     patch_v_s_route_from_yaml
 from suite.resources_utils import delete_service, get_first_pod_name, get_events, \
