+++ tests/suite/test_v_s_route_canned_responses.py
@@ -6,7 +6,7 @@ from kubernetes.client.rest import ApiException
 from settings import TEST_DATA
 from suite.custom_assertions import assert_event_and_get_count, wait_and_assert_status_code, \
     assert_event_count_increased, assert_event_starts_with_text_and_contains_errors
-from suite.custom_resources_utils import get_vs_nginx_template_conf, patch_v_s_route_from_yaml
+from suite.vs_vsr_resources_utils import get_vs_nginx_template_conf, patch_v_s_route_from_yaml
 from suite.resources_utils import get_first_pod_name, get_events, wait_before_test
