+++ tests/suite/test_v_s_route_focused_canary.py
@@ -3,7 +3,7 @@ import requests
 import yaml
 
 from settings import TEST_DATA
-from suite.custom_resources_utils import create_virtual_server_from_yaml, create_v_s_route_from_yaml
+from suite.vs_vsr_resources_utils import create_virtual_server_from_yaml, create_v_s_route_from_yaml
 from suite.fixtures import VirtualServerRoute
 from suite.resources_utils import ensure_response_from_backend, create_example_app, \
     wait_until_all_pods_are_ready, create_namespace_with_name_from_yaml, delete_namespace
