+++ tests/suite/test_virtual_server_status.py
@@ -1,9 +1,11 @@
 import pytest
 from kubernetes.client.rest import ApiException
 from suite.resources_utils import wait_before_test
+from suite.vs_vsr_resources_utils import (
+    patch_virtual_server_from_yaml,
+)
 from suite.custom_resources_utils import (
     read_custom_resource,
-    patch_virtual_server_from_yaml,
 )
 from settings import TEST_DATA
